requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
python-dotenv
google-genai
requests
Pillow
google-generativeai
